+++ pom.xml
@@ -37,8 +37,8 @@
 	</licenses>
 
 	<properties>
-		<maven.compiler.target>1.8</maven.compiler.target>
-		<maven.compiler.source>1.8</maven.compiler.source>
+		<maven.compiler.target>11</maven.compiler.target>
+		<maven.compiler.source>11</maven.compiler.source>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<com.github.ben-manes.caffeine.caffeine.version>2.9.3</com.github.ben-manes.caffeine.caffeine.version>
 	</properties>
